Move tests out of distribution and use setuptools to run the tests.
We don't need the alltests.py file, since we can run the tests with
python setup.py test
