+++ test/activeresource_test.py
@@ -12,7 +12,7 @@ from pyactiveresource import activeresource
 from pyactiveresource import connection
 from pyactiveresource import formats
 from pyactiveresource import util
-from pyactiveresource.tests import http_fake
+from pyactiveresource.testing import http_fake
 
 
 class Error(Exception):
